portal_project/requirements.txt
# requirements.txt - ПРАВИЛЬНАЯ ВЕРСИЯ

streamlit
gspread
pandas
requests
streamlit-lottie
google-auth-oauthlib
oauth2client